move score_package to worker.scoring
