+++ setup.cfg
@@ -27,7 +27,8 @@ files =
       depobs/scanner/docker/*.py,
       depobs/scanner/models/*.py,
       depobs/scanner/pipelines/__init__.py,
-      depobs/website/models.py
+      depobs/website/models.py,
+      depobs/worker/scoring.py
 
 python_version = 3.8
 # compose mounts . into /app and mypy can't write to its default .mypy_cache dir
@@ -48,14 +49,21 @@ warn_unused_configs = True
 
 [mypy-depobs.website.models]
 check_untyped_defs = True
-disallow_incomplete_defs = True
-disallow_untyped_calls = True
-disallow_untyped_decorators = True
 disallow_untyped_defs = True
+disallow_untyped_calls = True
+disallow_incomplete_defs = True
+disallow_untyped_decorators = True
+
+[mypy-depobs.worker.scoring]
+check_untyped_defs = True
+disallow_untyped_defs = True
+disallow_untyped_calls = True
+disallow_incomplete_defs = True
+disallow_untyped_decorators = True
 
 [mypy-depobs.scanner.graph_traversal]
 check_untyped_defs = True
-disallow_incomplete_defs = True
-disallow_untyped_calls = True
-disallow_untyped_decorators = True
 disallow_untyped_defs = True
+disallow_untyped_calls = True
+disallow_incomplete_defs = True
+disallow_untyped_decorators = True
